ml_api/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
gunicorn
python-dotenv